+++ lib/change-controller.ts
@@ -1,8 +1,8 @@
 import { aws_cloudwatch as cloudwatch, aws_codepipeline as cp, aws_events as
   events, aws_events_targets as events_targets, aws_iam as iam, aws_lambda as
   lambda, aws_s3 as s3, aws_s3_notifications as s3_notifications }
-  from "monocdk-experiment";
-  import * as cdk from 'monocdk-experiment';
+  from "monocdk";
+  import * as cdk from 'monocdk';
 import path = require("path");
 
 export interface ChangeControllerProps {
@@ -70,7 +70,15 @@ export class ChangeController extends cdk.Construct {
 
     const fn = new lambda.Function(this, 'Function', {
       description: `Enforces a Change Control Policy into CodePipeline's ${props.pipelineStage.stageName} stage`,
-      code: lambda.Code.fromAsset(path.join(__dirname, '../change-control-lambda'), { exclude: ['*.tsbuildinfo'] }),
+      code: lambda.Code.fromAsset(path.join(__dirname, '../change-control-lambda'), {
+        exclude: [
+          '**/.DS_Store',
+          '*.tsbuildinfo',
+          '*.ts',
+          'tsconfig.json',
+          'yarn.lock',
+        ],
+      }),
       runtime: lambda.Runtime.NODEJS_10_X,
       handler: 'index.handler',
       environment: {